unify SRS data and audio clients
